Did a buncha review stuff, implemented accent, moved to _wizden filestructure
